+++ starkbank/brcodepayment/__init__.py
@@ -1,3 +1,4 @@
 from .__brcodepayment import create, get, query, page, pdf, update
 from .log.__log import Log
 from . import log
+from .rule.__rule import Rule