+++ tests/api/testTransaction.py
@@ -91,3 +91,4 @@ class TestTransactionInfoGet(TestCase):
 
 if __name__ == '__main__':
     main()
+